Add: 커미션 상세페이지 팔로우 버튼 api 연결, 작가정보 프로필 이미지 연동

+++ app/src/main/java/com/example/commit/connection/dto/CommissionArtistResponse.kt
@@ -9,7 +9,7 @@ data class CommissionArtistResponse(
 )
 
 data class ArtistSummary(
-    val artistId: Long,
+    val artistId: Int,
     val nickname: String,
     val profileImageUrl: String?,
     val follower: Int,
